Handling of pre-epoch dates on Windows.
On Windows (and perhaps other platforms as well), pre-UNIX-epoch dates, expressed as negative seconds since the epoch, lead to ValueErrors. Since all we do really is pick one of two local timezones based on the DST flag, we can fall back to using the date with the year swapped out for the current year instead.